Update TODO and comment on CSS
As of 2022 there is no way to observe custom CSS property changes.
This cripples custom WebComponents from the get-go. A nice design should
include the ability to recreate built-in widgets as a priority goal.

Right now we have magical built-in elements and half-assed user
elements.

Additional rant:

There is no simple way to read a custom element text content other than
using a <slot> whatever that overengineered PoS stands for.

https://stackoverflow.com/questions/48498581/textcontent-empty-in-connectedcallback-of-a-custom-htmlelement